+++ CHANGELOG.md
@@ -1,13 +1,27 @@
 # Changelog
 
 
-## [0.0.0] - 2025-01-01
+## [0.1.2] - 2025-01-08
 ### Add
+- terminate topic
+- remembr_status topic
+- goal_text topic
 ### Change
+- use (x, y, z) in prompt example
+- fix string input for retrieve_from_position
+- split AIMessage
+- new query interrupt previous one
+- break infinite generate loop after three retries
 
-## [0.1.0] - 2024-12-28
+## [0.1.1] - 2024-12-20
+### Added
+### Change
+- test on amr
+- fix caption pose timestamp
+
+## [0.1.0] - 2024-12-18
 ### Added
 - webcam and fake pose node
-- fix common_utils bug
 - add caption pose publisher in caption node
 ### Change
+- fix common_utils bug